Cambios tarea-3

+++ tareas/weather-app/src/main/java/tec/bd/weather/service/WeatherServiceImpl.java
@@ -55,6 +55,10 @@ public class WeatherServiceImpl implements WeatherService{
     
     @Override
     public void removeForecast(int forecastId) {
+        var current = this.weatherRepository.findById(forecastId);
+        if (current.isEmpty()){
+            throw new RuntimeException("Weather forecast ID doesn't exists in database");
+        }
         this.weatherRepository.delete(forecastId);
     }
